+++ Weather Varta/Model/WeatherResult.swift
@@ -28,8 +28,8 @@ struct Temperature: Codable {
     var feels_like: Double
     var temp_min: Double
     var temp_max: Double
-    var pressure: Double
-    var humidity: Double
+    var pressure: Int
+    var humidity: Int
 }
 
 struct Wind: Codable {
